Refactor: foreach_{node,ticket}: apply more universally
This is a primitive that's easier to follow/more terse.
Looking forward, we skip some places where we want to introduce
a refactoring to drop an undesired non-transparent reliance on
booth_conf global variable prior to modifying these two macros
to that effect as well (respective places, at least, receive
white-space-around-operators sanitization).

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>